Modify codes for trace
1. Use Accept() to contain accept()

In echo.h , Accept() will address the return value from accept().
If accept() return error value, Accept() will call exit(0).It is for
setting a probe use uretprobe in bpftrace because I don't know how to
set a probe to accept().

2. Make client keeping send messages

In echoclient.c , replace fgets() with sprintf() to make client be able
to automatically send messages to server, which is for tracing easily.